requirements.txt

# Runtime dependencies
binance-connector>=3.12.0
websocket-client>=1.6.3
requests>=2.31.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.1.0

# Code quality
black>=23.10.1
ruff>=0.1.3
mypy>=1.6.1
types-requests>=2.31.0
